Integrated and robust deconvolution

This function provides a robust devonvolution framework to integrate information from scRNA-seq references, marker genes, and prior biological knowledge.

InteRD2(bulk.data,list_marker,cell_type_unique,comb_sampled,ave_est,ave_sd,lambda_option)

Arguments

bulk.data

ExpressionSet object for a target bulk data.

list_marker

A list of pre-specified marker genes corresponding to each cell type.

cell_type_unique

A list of cell types. It should match the order in list.marker.

comb_sampled

A pre-specified cell type proportions for the target bulk data, which could be obtained from reference-based deconvolution approach.

ave_est

A pre-specified population-level cell type proportions, which could be obtained from single-cell RNA-seq and external expression data from different studies, species, or data types

ave_sd

A pre-specified standard deviation for cell-type proportion estimation. The default is 1 for each cell type.

lambda_option

a sequence of values for the tunning parameter.

Value

A list containing estimated cell type proportions corresponding to each tuning value, named est; and a sequence of goodness-of-fit values corresponding to each tuning value, named metrics. The smaller the better.